pyproject.toml
[tool.poetry]
name = "hackathon-ai-for-good-2023-06"
version = "0.1.0"
description = ""
authors = ["Maciej J Mikulski <mjmikulski@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "hackathon_ai_for_good_2023_06"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.23.1"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
pandas = "^2.0.2"
opencv-python = "^4.7.0.72"
ipython = "^8.14.0"
notebook = "^6.5.4"
tiktoken = "0.3.3"
langchain = "0.0.142"
chromadb = "^0.3.26"
streamlit-chat = "^0.0.2.2"
faiss-cpu = "^1.7.4"
unstructured = "^0.7.5"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
